Refactoring for modularity
